kubernetes/spectrogram/requirements.txt
google-cloud-pubsub==0.41.0
google-cloud-storage==1.16.0
librosa==0.6.3
pypng==0.0.19
Pillow==6.0.0
python-json-logger==0.1.11
